Fix Codex review: invite race + Postgres bigint count cast
P1 — atomic invite consumption (src/routes/join/[code], src/routes/signup):
The previous flow read the invitation, then unconditionally INSERTed a
membership and UPDATEd the invite. Two concurrent claims from different
users could both pass the read and both end up with access from a
one-shot code (the (user_id, child_id) PK only protects against the
*same* user double-claiming). Wrap the consumption in a transaction
with `WHERE invitations.code = ? AND used_at IS NULL` and check
rowCount; on the join route we abort the tx and return the
already-expired error, on signup we throw a sentinel that rolls back
the entire user insert (no phantom account left behind). Added tests
that vi.spyOn `testDb.transaction` to mark the invite consumed
mid-flow and assert the race branches.

P2 — count(*) cast (src/routes/child/[id]/log/+page.server.ts):
node-postgres returns int8 columns as strings; drizzle's `sql<number>`
is only a TypeScript hint. The milestone redirects compared
`priorEntryCount === 0`, `priorAllergenCount === 0` against the string
'0', so first-food / first-allergen / all-12-allergens query params
never fired in production. Changed every count projection to
`count(*)::int` / `count(distinct …)::int`. Tests didn't catch this
because pg-mem returns counts as numbers natively — a documented
divergence we now have to budget around at the SQL layer.

+++ src/routes/child/[id]/log/+page.server.ts
@@ -132,11 +132,15 @@ export const actions: Actions = {
             throw new LogActionAbort(400, 'Aliment introuvable.');
           }
 
-          // Snapshot pre-insert state so we can detect milestones after the insert.
+          // Snapshot pre-insert state so we can detect milestones after the
+          // insert. Cast counts to int — node-postgres returns BIGINT as a
+          // string by default, which would make every `=== 0` comparison
+          // below false and silently disable the first-food / first-allergen
+          // / all-allergens redirects.
           const priorEntryCount =
             (
               await tx
-                .select({ n: sql<number>`count(*)` })
+                .select({ n: sql<number>`count(*)::int` })
                 .from(foodEntries)
                 .where(eq(foodEntries.childId, childId))
                 .limit(1)
@@ -147,7 +151,7 @@ export const actions: Actions = {
           const priorCategoriesCovered =
             (
               await tx
-                .select({ n: sql<number>`count(distinct ${foods.category})` })
+                .select({ n: sql<number>`count(distinct ${foods.category})::int` })
                 .from(foodEntries)
                 .innerJoin(foods, eq(foods.id, foodEntries.foodId))
                 .where(and(eq(foodEntries.childId, childId), ne(foods.category, 'autre')))
@@ -158,7 +162,7 @@ export const actions: Actions = {
             food.allergenType != null
               ? ((
                   await tx
-                    .select({ n: sql<number>`count(*)` })
+                    .select({ n: sql<number>`count(*)::int` })
                     .from(foodEntries)
                     .innerJoin(foods, eq(foods.id, foodEntries.foodId))
                     .where(
@@ -176,7 +180,7 @@ export const actions: Actions = {
           const priorAllergensIntroduced =
             (
               await tx
-                .select({ n: sql<number>`count(distinct ${foods.allergenType})` })
+                .select({ n: sql<number>`count(distinct ${foods.allergenType})::int` })
                 .from(foodEntries)
                 .innerJoin(foods, eq(foods.id, foodEntries.foodId))
                 .where(
@@ -198,7 +202,7 @@ export const actions: Actions = {
           const categoriesNowCovered =
             (
               await tx
-                .select({ n: sql<number>`count(distinct ${foods.category})` })
+                .select({ n: sql<number>`count(distinct ${foods.category})::int` })
                 .from(foodEntries)
                 .innerJoin(foods, eq(foods.id, foodEntries.foodId))
                 .where(and(eq(foodEntries.childId, childId), ne(foods.category, 'autre')))
